Moved the pretty printed verification report to VerificationReport.scala
